+++ dapp/src/app/globals.css
@@ -4,8 +4,6 @@
 @tailwind components;
 @tailwind utilities;
 
-/* @custom-variant dark (&:where(.dark, .dark *)); */
-
 :root {
     @apply bg-neutral-100 dark:bg-neutral-900;
     @apply text-neutral-900 dark:text-neutral-100;
