+++ dapp/src/contexts/index.ts
@@ -1,1 +1,4 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 export * from './ThemeContext';
