+++ dapp/src/lib/constants/index.ts
@@ -3,3 +3,4 @@
 
 export * from './errorMessages.constants';
 export * from './token';
+export * from './routes.constants';
