+++ dapp/src/lib/enums/protectedRouteTitle.enums.ts
@@ -1,5 +1,6 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './theme.enums';
-export * from './protectedRouteTitle.enums';
+export enum ProtectedRouteTitle {
+    MyNames = 'My Names',
+}
