+++ dapp/src/lib/enums/theme.enums.ts
@@ -1,3 +1,6 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 export enum Theme {
     Light = 'light',
     Dark = 'dark',
