+++ dapp/src/providers/ThemeProvider.tsx
@@ -1,3 +1,6 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 'use client';
 
 import { useCallback, useEffect, useState } from 'react';
